Abstracted stochastic MCMC kernel parts from SGHMC
